chore: Modularize codebase, fix bugs, and enhance core features
- Split app.py into flow.py, ui.py, actions.py, utils.py for better modularity.
- Added unique hat_id generation to prevent conflicts.
- Fixed Ollama JSON output validation with retry logic.
- Fixed Chainlit Message.update() async bug.
- Fixed schedule-based hat switching to properly update sessions.
- Enhanced Hat schema with critics, qa_loop, memory_tags, flow_order, etc.
- Enabled Critic QA loop for any Hat via dynamic critics[] logic.
- Added timestamp-based dynamic team_id in team creation.
- Show current Hat JSON on 'edit <hat_id>' for easier updates.
- Validated full team flow (summarizer + critic + custom).
- Polished memory formatting and debug utilities.
- Finalized all 7 test cases including sidebar, team flow, and scheduling.
