+++ app.py
@@ -7,12 +7,39 @@ import re
 
 from datetime import datetime
 
-from hat_manager import get_vector_db_for_hat, list_hats_by_team, search_memory, add_memory_to_hat, clear_memory
+from hat_manager import (
+    list_hats,
+    list_hats_by_team,
+    load_hat,
+    save_hat,
+    create_hat_from_prompt,
+    ollama_llm,
+    get_vector_db_for_hat,
+    search_memory,
+    add_memory_to_hat,
+    clear_memory
+)
 
 import openai
 
 import os
 from dotenv import load_dotenv
+
+from ui import show_hat_sidebar, show_hat_selector
+from actions import (
+    ask_for_prompt,
+    handle_prompt,
+    save_edited_json,
+    wear_hat_action_button,
+    edit_hat_action_ui,
+    save_hat_action_ui,
+    save_schedule_action,
+    handle_hat_selection
+)
+
+from flow import run_team_flow
+
+from utils import generate_unique_hat_id, current_timestamp, format_memory_entry
 
 load_dotenv()
 # Set your OpenAI API key
@@ -55,304 +82,304 @@ async def main():
     await show_hat_selector()
     await cl.Message(content="👋 Welcome! Select a Hat, use commands, or type `help`.").send()
 
-# --- UI Update Functions ---
+# # --- UI Update Functions ---
 
-async def show_hat_sidebar():
-    hat_ids = list_hats()
-    hats = [load_hat(hat_id) for hat_id in hat_ids]  # Load full Hat data
+# async def show_hat_sidebar():
+#     hat_ids = list_hats()
+#     hats = [load_hat(hat_id) for hat_id in hat_ids]  # Load full Hat data
 
-    team_ids = list(set([hat.get("team_id") for hat in hats if hat.get("team_id")]))
+#     team_ids = list(set([hat.get("team_id") for hat in hats if hat.get("team_id")]))
     
-    elements = [Text(content="### 🎩 Hats")]
+#     elements = [Text(content="### 🎩 Hats")]
 
-    if hat_ids:
-        for hat_id in hat_ids:
-            elements.append(Text(content=f"- `wear {hat_id}`"))
-    else:
-        elements.append(Text(content="_(No hats found yet)_"))
+#     if hat_ids:
+#         for hat_id in hat_ids:
+#             elements.append(Text(content=f"- `wear {hat_id}`"))
+#     else:
+#         elements.append(Text(content="_(No hats found yet)_"))
 
-    elements += [
-        Text(content="---"),
-        Text(content="### ➕ Create"),
-        Text(content="- `new blank` — Empty Hat"),
-        Text(content="- `new from prompt` — Describe"),
-        Text(content="- `create team` — Auto-build a team"),
-        Text(content="---"),
-        Text(content="### 👥 Teams"),
-        Text(content="- `run team <team_id>` — Run multi-Hat flow"),
-        Text(content="- `view team <team_id>` — Show team Hats"),
-        Text(content="- `save team` — Save proposed team"),
-        Text(content="- `show team json` — View team JSON"),
-        Text(content="---"),
-        Text(content="### 🧠 Memory"),
-        Text(content="- `view memories` — Show top memories"),
-        Text(content="- `clear memories` — Delete all memories"),
-        Text(content="- `export memories <hat_id>` — Export JSON"),
-        Text(content="- `debug memories` — Raw memory debug"),
-        Text(content="---"),
-        Text(content="### 🕒 Scheduling"),
-        Text(content="- `set schedule` — Schedule Hat switching"),
-        Text(content="- `view schedule` — View schedules"),
-        Text(content="- `current hat` — Show active Hat"),
-        Text(content="---"),
-        Text(content="### ✅ Active Teams")
-    ]
+#     elements += [
+#         Text(content="---"),
+#         Text(content="### ➕ Create"),
+#         Text(content="- `new blank` — Empty Hat"),
+#         Text(content="- `new from prompt` — Describe"),
+#         Text(content="- `create team` — Auto-build a team"),
+#         Text(content="---"),
+#         Text(content="### 👥 Teams"),
+#         Text(content="- `run team <team_id>` — Run multi-Hat flow"),
+#         Text(content="- `view team <team_id>` — Show team Hats"),
+#         Text(content="- `save team` — Save proposed team"),
+#         Text(content="- `show team json` — View team JSON"),
+#         Text(content="---"),
+#         Text(content="### 🧠 Memory"),
+#         Text(content="- `view memories` — Show top memories"),
+#         Text(content="- `clear memories` — Delete all memories"),
+#         Text(content="- `export memories <hat_id>` — Export JSON"),
+#         Text(content="- `debug memories` — Raw memory debug"),
+#         Text(content="---"),
+#         Text(content="### 🕒 Scheduling"),
+#         Text(content="- `set schedule` — Schedule Hat switching"),
+#         Text(content="- `view schedule` — View schedules"),
+#         Text(content="- `current hat` — Show active Hat"),
+#         Text(content="---"),
+#         Text(content="### ✅ Active Teams")
+#     ]
 
-    for team_id in team_ids:
-        team_hats = list_hats_by_team(team_id)
-        hat_names = ", ".join([hat.get("name", "Unnamed") for hat in team_hats])
-        elements.append(Text(content=f"- `{team_id}`: {hat_names}"))
+#     for team_id in team_ids:
+#         team_hats = list_hats_by_team(team_id)
+#         hat_names = ", ".join([hat.get("name", "Unnamed") for hat in team_hats])
+#         elements.append(Text(content=f"- `{team_id}`: {hat_names}"))
 
-    elements += [
-        Text(content="---"),
-        Text(content="### 🔥 TODOs / Next Steps"),
-        Text(content="- `user approval` after Critic — Await 'approve'/'retry' ✔️"),
-        Text(content="- Visual polish for logs (emojis/dividers) ✔️"),
-        Text(content="- `import memories` command — #TODO"),
-        Text(content="- Add `run again` / `create new team` buttons — #TODO"),
-        Text(content="- QA edge fallback: prompt user after retry limit — #TODO"),
-        Text(content="- Flow visualizer: plan graphical flow editor — #TODO"),
-    ]
+#     elements += [
+#         Text(content="---"),
+#         Text(content="### 🔥 TODOs / Next Steps"),
+#         Text(content="- `user approval` after Critic — Await 'approve'/'retry' ✔️"),
+#         Text(content="- Visual polish for logs (emojis/dividers) ✔️"),
+#         Text(content="- `import memories` command — #TODO"),
+#         Text(content="- Add `run again` / `create new team` buttons — #TODO"),
+#         Text(content="- QA edge fallback: prompt user after retry limit — #TODO"),
+#         Text(content="- Flow visualizer: plan graphical flow editor — #TODO"),
+#     ]
 
-    await cl.ElementSidebar.set_elements(elements)
-    await cl.ElementSidebar.set_title("Hat Manager")
+#     await cl.ElementSidebar.set_elements(elements)
+#     await cl.ElementSidebar.set_title("Hat Manager")
 
-async def show_hat_selector():
-    """
-    Displays dynamic action buttons in chat to wear hats.
-    """
-    hats = list_hats()
-    if not hats:
-        return
+# async def show_hat_selector():
+#     """
+#     Displays dynamic action buttons in chat to wear hats.
+#     """
+#     hats = list_hats()
+#     if not hats:
+#         return
 
-    # Create actions, putting the hat_id into the payload
-    actions = [
-        Action(
-            name="wear_hat_button",
-            label=hat,                 # Text displayed on the button
-            # value=hat,               # Remove or comment out - not used in callback
-            payload={"hat_id": hat}    # Put the hat ID into the payload dictionary
-        ) for hat in hats
-    ]
+#     # Create actions, putting the hat_id into the payload
+#     actions = [
+#         Action(
+#             name="wear_hat_button",
+#             label=hat,                 # Text displayed on the button
+#             # value=hat,               # Remove or comment out - not used in callback
+#             payload={"hat_id": hat}    # Put the hat ID into the payload dictionary
+#         ) for hat in hats
+#     ]
 
-    await cl.Message(
-        content="🎩 Select a Hat to wear:",
-        actions=actions
-    ).send()
-# --- Action Callbacks (Button Clicks & Custom UI Events) ---
+#     await cl.Message(
+#         content="🎩 Select a Hat to wear:",
+#         actions=actions
+#     ).send()
+# # --- Action Callbacks (Button Clicks & Custom UI Events) ---
 
-@cl.action_callback("wear_hat_button")
-async def wear_hat_action_button(action: Action):
-    """Handles clicks on the 'wear hat' buttons."""
-    # Retrieve hat_id from the payload dictionary using .get for safety
-    hat_id = action.payload.get("hat_id")
+# @cl.action_callback("wear_hat_button")
+# async def wear_hat_action_button(action: Action):
+#     """Handles clicks on the 'wear hat' buttons."""
+#     # Retrieve hat_id from the payload dictionary using .get for safety
+#     hat_id = action.payload.get("hat_id")
 
-    if not hat_id:
-        # Handle case where payload might be missing the key
-        await cl.Message(content="❌ Error: Could not retrieve hat_id from action payload.").send()
-        return
+#     if not hat_id:
+#         # Handle case where payload might be missing the key
+#         await cl.Message(content="❌ Error: Could not retrieve hat_id from action payload.").send()
+#         return
 
-    # hat_id = action.value # <<< This line caused the AttributeError
+#     # hat_id = action.value # <<< This line caused the AttributeError
 
-    # Proceed with the retrieved hat_id
-    await wear_hat(hat_id)
+#     # Proceed with the retrieved hat_id
+#     await wear_hat(hat_id)
 
-@cl.action_callback("edit_hat_ui")
-async def edit_hat_action_ui(action: Action):
-    """Handles click on the 'Edit This Hat' button, showing the custom UI."""
-    hat_id = action.payload.get("hat_id")
-    if not hat_id:
-        await cl.Message(content="❌ Error: Missing hat_id in edit action payload.").send()
-        return
+# @cl.action_callback("edit_hat_ui")
+# async def edit_hat_action_ui(action: Action):
+#     """Handles click on the 'Edit This Hat' button, showing the custom UI."""
+#     hat_id = action.payload.get("hat_id")
+#     if not hat_id:
+#         await cl.Message(content="❌ Error: Missing hat_id in edit action payload.").send()
+#         return
 
-    try:
-        hat = load_hat(hat_id)
-        if hat is None:
-            await cl.Message(content=f"❌ Error: Could not load hat with ID '{hat_id}'.").send()
-            return
+#     try:
+#         hat = load_hat(hat_id)
+#         if hat is None:
+#             await cl.Message(content=f"❌ Error: Could not load hat with ID '{hat_id}'.").send()
+#             return
 
-        print(f"Hat data being passed as props (JSON encoded): {json.dumps(hat)}")
+#         print(f"Hat data being passed as props (JSON encoded): {json.dumps(hat)}")
 
-        await cl.Message(
-            content="Testing basic HatEditor:",
-            elements=[
-                cl.CustomElement(
-                    name="TestComponent",
-                    display="inline",
-                    props=hat
-                )
-            ]
-        ).send()
-    except Exception as e:
-        await cl.Message(content=f"❌ Error loading hat '{hat_id}' for UI editing: {e}").send()
+#         await cl.Message(
+#             content="Testing basic HatEditor:",
+#             elements=[
+#                 cl.CustomElement(
+#                     name="TestComponent",
+#                     display="inline",
+#                     props=hat
+#                 )
+#             ]
+#         ).send()
+#     except Exception as e:
+#         await cl.Message(content=f"❌ Error loading hat '{hat_id}' for UI editing: {e}").send()
 
 
-@cl.action_callback("save_hat")
-async def save_hat_action_ui(action: Action):
-    """
-    Handles the 'save_hat' action emitted by the custom HatEditor UI component.
-    """
-    updated_hat = action.payload # Assumes payload is the complete updated hat dictionary
-    print(f"Received updated hat from UI: {updated_hat}")
-    hat_id = updated_hat.get("hat_id")
+# @cl.action_callback("save_hat")
+# async def save_hat_action_ui(action: Action):
+#     """
+#     Handles the 'save_hat' action emitted by the custom HatEditor UI component.
+#     """
+#     updated_hat = action.payload # Assumes payload is the complete updated hat dictionary
+#     print(f"Received updated hat from UI: {updated_hat}")
+#     hat_id = updated_hat.get("hat_id")
 
-    if not hat_id or not isinstance(updated_hat, dict):
-        await cl.Message(content="❌ Error: Invalid payload received from Hat Editor.").send()
-        return
+#     if not hat_id or not isinstance(updated_hat, dict):
+#         await cl.Message(content="❌ Error: Invalid payload received from Hat Editor.").send()
+#         return
 
-    try:
-        save_hat(hat_id, updated_hat)
-        cl.user_session.set("current_hat", updated_hat) # Update session
-        cl.user_session.set("editing_hat_id", hat_id) # Keep track of active hat
-        await cl.Message(content=f"✅ Hat '{updated_hat.get('name', hat_id)}' updated via UI.").send()
-        await cl.Message(
-    content="✅ Hat saved. Here’s the updated version for further edits:",
-    elements=[
-        cl.CustomElement(name="TestComponent", id="hat_editor", props=updated_hat, display="inline")
-    ]
-).send()
-        await show_hat_sidebar() # Refresh sidebar
-        await show_hat_selector() # Refresh buttons
-    except Exception as e:
-        await cl.Message(content=f"❌ Error saving hat '{hat_id}' from UI: {e}").send()
+#     try:
+#         save_hat(hat_id, updated_hat)
+#         cl.user_session.set("current_hat", updated_hat) # Update session
+#         cl.user_session.set("editing_hat_id", hat_id) # Keep track of active hat
+#         await cl.Message(content=f"✅ Hat '{updated_hat.get('name', hat_id)}' updated via UI.").send()
+#         await cl.Message(
+#     content="✅ Hat saved. Here’s the updated version for further edits:",
+#     elements=[
+#         cl.CustomElement(name="TestComponent", id="hat_editor", props=updated_hat, display="inline")
+#     ]
+# ).send()
+#         await show_hat_sidebar() # Refresh sidebar
+#         await show_hat_selector() # Refresh buttons
+#     except Exception as e:
+#         await cl.Message(content=f"❌ Error saving hat '{hat_id}' from UI: {e}").send()
         
-@cl.action_callback("save_schedule")
-async def save_schedule_action(action: cl.Action):
-    time_str = action.inputs.get("schedule_time")
-    hat_id = action.inputs.get("schedule_hat")
+# @cl.action_callback("save_schedule")
+# async def save_schedule_action(action: cl.Action):
+#     time_str = action.inputs.get("schedule_time")
+#     hat_id = action.inputs.get("schedule_hat")
 
-    if time_str and hat_id:
-        schedule = cl.user_session.get("hat_schedule", {})
-        schedule[time_str] = hat_id
-        cl.user_session.set("hat_schedule", schedule)
+#     if time_str and hat_id:
+#         schedule = cl.user_session.get("hat_schedule", {})
+#         schedule[time_str] = hat_id
+#         cl.user_session.set("hat_schedule", schedule)
 
-        await cl.Message(content=f"✅ Scheduled Hat `{hat_id}` for `{time_str}`!").send()
-    else:
-        await cl.Message(content="❌ Please provide both a time and a hat.").send()
+#         await cl.Message(content=f"✅ Scheduled Hat `{hat_id}` for `{time_str}`!").send()
+#     else:
+#         await cl.Message(content="❌ Please provide both a time and a hat.").send()
 
-@cl.action_callback("hat_selection")
-async def handle_hat_selection(action: cl.Action):
-    selected_hat = action.value
-    schedule_time = cl.user_session.get("schedule_time_temp")
+# @cl.action_callback("hat_selection")
+# async def handle_hat_selection(action: cl.Action):
+#     selected_hat = action.value
+#     schedule_time = cl.user_session.get("schedule_time_temp")
 
-    if selected_hat and schedule_time:
-        schedule = cl.user_session.get("hat_schedule", {})
-        schedule[schedule_time] = selected_hat
-        cl.user_session.set("hat_schedule", schedule)
+#     if selected_hat and schedule_time:
+#         schedule = cl.user_session.get("hat_schedule", {})
+#         schedule[schedule_time] = selected_hat
+#         cl.user_session.set("hat_schedule", schedule)
 
-        await cl.Message(content=f"✅ Scheduled Hat `{selected_hat}` for `{schedule_time}`!").send()
-    else:
-        await cl.Message(content="❌ Failed to schedule. Try again.").send()
+#         await cl.Message(content=f"✅ Scheduled Hat `{selected_hat}` for `{schedule_time}`!").send()
+#     else:
+#         await cl.Message(content="❌ Failed to schedule. Try again.").send()
 
 
 # --- Core Logic Functions ---
 
-#Team building
-async def run_team_flow(team_id, input_text):
-    team_hats = list_hats_by_team(team_id)
-    current_input = input_text
-    conversation_log = []
-    retry_counts = {}
+# #Team building
+# async def run_team_flow(team_id, input_text):
+#     team_hats = list_hats_by_team(team_id)
+#     current_input = input_text
+#     conversation_log = []
+#     retry_counts = {}
 
-    i = 0
-    while i < len(team_hats):
-        hat = team_hats[i]
-        hat_name = hat.get('name')
-        hat_id = hat.get('hat_id')
-        instructions = hat.get('instructions')
-        qa_loop = hat.get('qa_loop', False)
-        retry_limit = hat.get('retry_limit', 0)
+#     i = 0
+#     while i < len(team_hats):
+#         hat = team_hats[i]
+#         hat_name = hat.get('name')
+#         hat_id = hat.get('hat_id')
+#         instructions = hat.get('instructions')
+#         qa_loop = hat.get('qa_loop', False)
+#         retry_limit = hat.get('retry_limit', 0)
 
-        response_text = generate_openai_response(current_input, hat)
+#         response_text = generate_openai_response(current_input, hat)
 
-        # Save memories
-        add_memory_to_hat(hat_id, current_input, role="user")
-        add_memory_to_hat(hat_id, response_text, role="bot")
+#         # Save memories
+#         add_memory_to_hat(hat_id, current_input, role="user")
+#         add_memory_to_hat(hat_id, response_text, role="bot")
 
-        await cl.Message(content=f"🧢 {hat_name} responded:\n{response_text}").send()
+#         await cl.Message(content=f"🧢 {hat_name} responded:\n{response_text}").send()
 
-        conversation_log.append({
-            "hat_name": hat_name,
-            "hat_id": hat_id,
-            "input": current_input,
-            "output": response_text
-        })
+#         conversation_log.append({
+#             "hat_name": hat_name,
+#             "hat_id": hat_id,
+#             "input": current_input,
+#             "output": response_text
+#         })
 
-        # --- QA Loop Check ---
-        if qa_loop:
-            if "#REVISION_REQUIRED" in response_text:
-                retry_target_index = len(conversation_log) - 2  # Retry previous Hat
-                retry_target = conversation_log[retry_target_index] if retry_target_index >= 0 else None
+#         # --- QA Loop Check ---
+#         if qa_loop:
+#             if "#REVISION_REQUIRED" in response_text:
+#                 retry_target_index = len(conversation_log) - 2  # Retry previous Hat
+#                 retry_target = conversation_log[retry_target_index] if retry_target_index >= 0 else None
 
-                if retry_target:
-                    retry_counts[retry_target['hat_id']] = retry_counts.get(retry_target['hat_id'], 0) + 1
+#                 if retry_target:
+#                     retry_counts[retry_target['hat_id']] = retry_counts.get(retry_target['hat_id'], 0) + 1
 
-                    if retry_counts[retry_target['hat_id']] <= retry_limit:
-                        await cl.Message(content=f"🔁 Critic requested revision. Retrying {retry_target['hat_name']}... (Attempt {retry_counts[retry_target['hat_id']]}/{retry_limit})").send()
+#                     if retry_counts[retry_target['hat_id']] <= retry_limit:
+#                         await cl.Message(content=f"🔁 Critic requested revision. Retrying {retry_target['hat_name']}... (Attempt {retry_counts[retry_target['hat_id']]}/{retry_limit})").send()
 
-                        # Re-run only the previous agent (e.g., Summarizer)
-                        prev_hat = next(h for h in team_hats if h['hat_id'] == retry_target['hat_id'])
-                        retry_response = generate_openai_response(retry_target['input'], prev_hat)
+#                         # Re-run only the previous agent (e.g., Summarizer)
+#                         prev_hat = next(h for h in team_hats if h['hat_id'] == retry_target['hat_id'])
+#                         retry_response = generate_openai_response(retry_target['input'], prev_hat)
 
-                        add_memory_to_hat(prev_hat['hat_id'], retry_target['input'], role="user")
-                        add_memory_to_hat(prev_hat['hat_id'], retry_response, role="bot")
+#                         add_memory_to_hat(prev_hat['hat_id'], retry_target['input'], role="user")
+#                         add_memory_to_hat(prev_hat['hat_id'], retry_response, role="bot")
 
-                        await cl.Message(content=f"🧢 {prev_hat['name']} retry responded:\n{retry_response}").send()
+#                         await cl.Message(content=f"🧢 {prev_hat['name']} retry responded:\n{retry_response}").send()
 
-                        # Immediately re-run Critic (this hat) again with new Summarizer output
-                        critic_response = generate_openai_response(retry_response, hat)
+#                         # Immediately re-run Critic (this hat) again with new Summarizer output
+#                         critic_response = generate_openai_response(retry_response, hat)
 
-                        add_memory_to_hat(hat['hat_id'], retry_response, role="user")
-                        add_memory_to_hat(hat['hat_id'], critic_response, role="bot")
+#                         add_memory_to_hat(hat['hat_id'], retry_response, role="user")
+#                         add_memory_to_hat(hat['hat_id'], critic_response, role="bot")
 
-                        await cl.Message(content=f"🧢 {hat_name} re-reviewed:\n{critic_response}").send()
+#                         await cl.Message(content=f"🧢 {hat_name} re-reviewed:\n{critic_response}").send()
 
-                        # Handle approval/revision after retry
-                        if "#REVISION_REQUIRED" in critic_response:
-                            continue  # Go back and retry again if needed
+#                         # Handle approval/revision after retry
+#                         if "#REVISION_REQUIRED" in critic_response:
+#                             continue  # Go back and retry again if needed
 
-                        elif "#APPROVED" in critic_response:
-                            await cl.Message(content="✅ Critic approved the output after retry!").send()
+#                         elif "#APPROVED" in critic_response:
+#                             await cl.Message(content="✅ Critic approved the output after retry!").send()
 
-                        else:
-                            await cl.Message(content="⚠️ Critic did not provide a valid tag. Manual review needed.").send()
+#                         else:
+#                             await cl.Message(content="⚠️ Critic did not provide a valid tag. Manual review needed.").send()
 
-                        # Ask user for final approval
-                        await cl.Message(content="🧑‍⚖️ Approve or Retry? Type `approve` or `retry`.").send()
-                        cl.user_session.set("awaiting_user_approval", True)
-                        cl.user_session.set("pending_critique_input", retry_response)
-                        cl.user_session.set("pending_team_id", team_id)
-                        return
+#                         # Ask user for final approval
+#                         await cl.Message(content="🧑‍⚖️ Approve or Retry? Type `approve` or `retry`.").send()
+#                         cl.user_session.set("awaiting_user_approval", True)
+#                         cl.user_session.set("pending_critique_input", retry_response)
+#                         cl.user_session.set("pending_team_id", team_id)
+#                         return
 
-                    else:
-                        await cl.Message(content="⚠️ Retry limit reached. Proceeding with available results.").send()
+#                     else:
+#                         await cl.Message(content="⚠️ Retry limit reached. Proceeding with available results.").send()
 
-                else:
-                    await cl.Message(content="⚠️ No valid previous Hat to retry.").send()
+#                 else:
+#                     await cl.Message(content="⚠️ No valid previous Hat to retry.").send()
 
-            elif "#APPROVED" in response_text:
-                await cl.Message(content="✅ Critic approved the output!").send()
+#             elif "#APPROVED" in response_text:
+#                 await cl.Message(content="✅ Critic approved the output!").send()
 
-            else:
-                await cl.Message(content="⚠️ Critic did not provide a valid tag. Manual review needed.").send()
+#             else:
+#                 await cl.Message(content="⚠️ Critic did not provide a valid tag. Manual review needed.").send()
 
-            # User approval after Critic (whether tagged or not)
-            await cl.Message(content="🧑‍⚖️ Approve or Retry? Type `approve` or `retry`.").send()
-            cl.user_session.set("awaiting_user_approval", True)
-            cl.user_session.set("pending_critique_input", current_input)
-            cl.user_session.set("pending_team_id", team_id)
-            return
+#             # User approval after Critic (whether tagged or not)
+#             await cl.Message(content="🧑‍⚖️ Approve or Retry? Type `approve` or `retry`.").send()
+#             cl.user_session.set("awaiting_user_approval", True)
+#             cl.user_session.set("pending_critique_input", current_input)
+#             cl.user_session.set("pending_team_id", team_id)
+#             return
 
-        current_input = response_text  # Pass to next Hat
-        i += 1
+#         current_input = response_text  # Pass to next Hat
+#         i += 1
 
-    # 📝 Full Log
-    log_text = "\n\n".join([
-        f"🧢 {entry['hat_name']}:\nInput: {entry['input']}\nOutput: {entry['output']}"
-        for entry in conversation_log
-    ])
-    await cl.Message(content=f"📜 **Full Conversation Log:**\n\n{log_text}").send()
-    await cl.Message(content="✅ Team flow complete!").send()
+#     # 📝 Full Log
+#     log_text = "\n\n".join([
+#         f"🧢 {entry['hat_name']}:\nInput: {entry['input']}\nOutput: {entry['output']}"
+#         for entry in conversation_log
+#     ])
+#     await cl.Message(content=f"📜 **Full Conversation Log:**\n\n{log_text}").send()
+#     await cl.Message(content="✅ Team flow complete!").send()
     
 async def wear_hat(hat_id: str):
     """Loads a hat, sets it as active in the session, and informs the user."""
@@ -389,23 +416,29 @@ async def wear_hat(hat_id: str):
         await cl.Message(content=f"❌ Couldn't load hat `{hat_id}`: {e}").send()
 
 async def create_blank_hat():
-    """Creates and saves a new, empty hat."""
-    # Find a unique ID (e.g., new_hat_1, new_hat_2)
-    i = 1
-    hat_id = "new_hat"
-    hats = list_hats()
-    while hat_id in hats:
-        hat_id = f"new_hat_{i}"
-        i += 1
+    """Creates and saves a new, empty hat with default enhanced schema."""
+    existing_ids = list_hats()  # Get current hat IDs
+    hat_id = generate_unique_hat_id(existing_ids)  # Pass them into the helper
+
 
     blank = {
         "hat_id": hat_id,
-        "name": f"New Hat {i-1 if i>1 else ''}".strip(),
-        "model": "default", # Or get from config
+        "name": f"New Hat {hat_id}",
+        "model": "gpt-3.5-turbo",  # Default model, you can change if needed
+        "role": "agent",
         "instructions": "",
         "tools": [],
-        "relationships": []
+        "relationships": [],
+        "team_id": None,
+        "flow_order": None,
+        "qa_loop": False,
+        "critics": [],
+        "active": True,
+        "memory_tags": [],
+        "retry_limit": 1,
+        "description": ""
     }
+
     try:
         save_hat(blank["hat_id"], blank)
         await cl.Message(content=f"📄 Blank Hat `{blank['hat_id']}` created. You can now `wear {blank['hat_id']}`.").send()
@@ -414,73 +447,73 @@ async def create_blank_hat():
     except Exception as e:
         await cl.Message(content=f"❌ Failed to save blank hat: {e}").send()
 
-async def ask_for_prompt():
-    """Prompts the user to describe the hat they want to create."""
-    cl.user_session.set("awaiting_hat_prompt", True)
-    cl.user_session.set("awaiting_json_paste", False) # Ensure other states are off
-    await cl.Message(content="🪄 Describe the new Hat you want to create (e.g., `A helpful assistant that can summarize text`).").send()
+# async def ask_for_prompt():
+#     """Prompts the user to describe the hat they want to create."""
+#     cl.user_session.set("awaiting_hat_prompt", True)
+#     cl.user_session.set("awaiting_json_paste", False) # Ensure other states are off
+#     await cl.Message(content="🪄 Describe the new Hat you want to create (e.g., `A helpful assistant that can summarize text`).").send()
 
-async def handle_prompt(message: cl.Message):
-    """Handles the user's description to create a hat using an LLM."""
-    cl.user_session.set("awaiting_hat_prompt", False) # Consume the flag
-    prompt_content = message.content
+# # async def handle_prompt(message: cl.Message):
+#     """Handles the user's description to create a hat using an LLM."""
+#     cl.user_session.set("awaiting_hat_prompt", False) # Consume the flag
+#     prompt_content = message.content
 
-    if not ollama_llm:
-         await cl.Message(content=f"❌ Cannot create from prompt: LLM (ollama_llm) is not configured.").send()
-         return
+#     if not ollama_llm:
+#          await cl.Message(content=f"❌ Cannot create from prompt: LLM (ollama_llm) is not configured.").send()
+#          return
 
-    try:
-        # Show thinking indicator
-        thinking_msg = cl.Message(content="⚙️ Creating Hat from your description...")
-        await thinking_msg.send()
+#     try:
+#         # Show thinking indicator
+#         thinking_msg = cl.Message(content="⚙️ Creating Hat from your description...")
+#         await thinking_msg.send()
 
-        hat = create_hat_from_prompt(prompt_content, ollama_llm) # External function call
+#         hat = create_hat_from_prompt(prompt_content, ollama_llm) # External function call
 
-        # Basic validation
-        if not hat or not hat.get("hat_id") or not hat.get("name"):
-             raise ValueError("LLM did not return a valid hat structure.")
+#         # Basic validation
+#         if not hat or not hat.get("hat_id") or not hat.get("name"):
+#              raise ValueError("LLM did not return a valid hat structure.")
 
-        save_hat(hat["hat_id"], hat)
+#         save_hat(hat["hat_id"], hat)
 
-        await thinking_msg.update(content=f"✅ Created Hat: `{hat['name']}` (ID: `{hat['hat_id']}`). You can now `wear {hat['hat_id']}`.")
-        await show_hat_sidebar()
-        await show_hat_selector()
-    except Exception as e:
-        await cl.Message(content=f"❌ Failed to create Hat from prompt: {e}").send()
+#         await thinking_msg.update(content=f"✅ Created Hat: `{hat['name']}` (ID: `{hat['hat_id']}`). You can now `wear {hat['hat_id']}`.")
+#         await show_hat_sidebar()
+#         await show_hat_selector()
+#     except Exception as e:
+#         await cl.Message(content=f"❌ Failed to create Hat from prompt: {e}").send()
 
 
-async def save_edited_json(message: cl.Message, hat_id_to_edit: str):
-    """Parses pasted JSON, validates, and saves the hat."""
-    try:
-        updated_hat = json.loads(message.content)
+# async def save_edited_json(message: cl.Message, hat_id_to_edit: str):
+#     """Parses pasted JSON, validates, and saves the hat."""
+#     try:
+#         updated_hat = json.loads(message.content)
 
-        if not isinstance(updated_hat, dict):
-            raise TypeError("Pasted content is not a valid JSON object.")
+#         if not isinstance(updated_hat, dict):
+#             raise TypeError("Pasted content is not a valid JSON object.")
 
-        # --- Validation ---
-        pasted_hat_id = updated_hat.get("hat_id")
-        if not pasted_hat_id:
-            raise ValueError("Pasted JSON is missing the 'hat_id' key.")
+#         # --- Validation ---
+#         pasted_hat_id = updated_hat.get("hat_id")
+#         if not pasted_hat_id:
+#             raise ValueError("Pasted JSON is missing the 'hat_id' key.")
 
-        if pasted_hat_id != hat_id_to_edit:
-            await cl.Message(content=f"❌ Mismatch: You are editing `{hat_id_to_edit}`, but the pasted JSON has `hat_id`: `{pasted_hat_id}`. Paste the correct JSON.").send()
-            return # Keep awaiting_json_paste as True
+#         if pasted_hat_id != hat_id_to_edit:
+#             await cl.Message(content=f"❌ Mismatch: You are editing `{hat_id_to_edit}`, but the pasted JSON has `hat_id`: `{pasted_hat_id}`. Paste the correct JSON.").send()
+#             return # Keep awaiting_json_paste as True
 
-        # --- Save ---
-        save_hat(hat_id_to_edit, updated_hat)
-        cl.user_session.set("current_hat", updated_hat) # Update active hat in session
-        # Keep editing_hat_id pointing to the current hat
-        await cl.Message(content=f"✅ Hat `{hat_id_to_edit}` updated from pasted JSON.").send()
-        await show_hat_sidebar() # Refresh UI
-        await show_hat_selector()
+#         # --- Save ---
+#         save_hat(hat_id_to_edit, updated_hat)
+#         cl.user_session.set("current_hat", updated_hat) # Update active hat in session
+#         # Keep editing_hat_id pointing to the current hat
+#         await cl.Message(content=f"✅ Hat `{hat_id_to_edit}` updated from pasted JSON.").send()
+#         await show_hat_sidebar() # Refresh UI
+#         await show_hat_selector()
 
-    except json.JSONDecodeError as e:
-        await cl.Message(content=f"❌ Invalid JSON: {e}. Please paste valid JSON content.").send()
-        # Keep awaiting_json_paste as True, let user try again
-    except Exception as e:
-        await cl.Message(content=f"❌ Error saving updated hat `{hat_id_to_edit}`: {e}").send()
-        # Optionally reset awaiting_json_paste here if the error is unrecoverable
-        # cl.user_session.set("awaiting_json_paste", False)
+#     except json.JSONDecodeError as e:
+#         await cl.Message(content=f"❌ Invalid JSON: {e}. Please paste valid JSON content.").send()
+#         # Keep awaiting_json_paste as True, let user try again
+#     except Exception as e:
+#         await cl.Message(content=f"❌ Error saving updated hat `{hat_id_to_edit}`: {e}").send()
+#         # Optionally reset awaiting_json_paste here if the error is unrecoverable
+#         # cl.user_session.set("awaiting_json_paste", False)
 
 
 def generate_openai_response(prompt: str, hat: dict):
@@ -529,35 +562,56 @@ Instructions: {instructions}.
     return response.choices[0].message.content
 
 def generate_team_from_goal(goal):
-    system_prompt = """
-    You are an AI that creates AI agent teams for complex tasks. Each agent (Hat) has a unique role.
-    
-    Based on the goal, define a team with:
-    - hat_id
-    - name
-    - role (planner, summarizer, critic, tool_agent)
-    - model (gpt-3.5 or gpt-4)
-    - instructions
-    - tools (list)
-    - flow_order (number)
-    - qa_loop (true/false)
-    - team_id ("auto_team")
 
-    Output valid JSON. Example:
+    # Dynamic team_id
+    team_id = f"auto_team_{datetime.now().strftime('%Y%m%d%H%M%S')}"
+
+    system_prompt = f"""
+    You are an AI that builds specialized multi-agent teams ("Hats") for complex tasks.
+
+    Each Hat must include:
+    - hat_id: short and unique (e.g., planner_{team_id})
+    - name: human-friendly name
+    - model: gpt-3.5-turbo
+    - role: planner, summarizer, critic, tool, researcher
+    - instructions: specific task details
+    - tools: [] or tool names
+    - relationships: [] or related hat_ids this agent collaborates with
+    - team_id: "{team_id}"
+    - flow_order: 1, 2, 3...
+    - qa_loop: true/false
+    - critics: hat_ids that review this hat
+    - active: true/false
+    - memory_tags: ["tag1", "tag2"]
+    - retry_limit: 1-3
+    - description: 1-line purpose
+
+    ### Goal:
+    {goal}
+
+    ### Example:
     [
-      {
-        "hat_id": "planner",
+      {{
+        "hat_id": "planner_{team_id}",
         "name": "Planning Agent",
+        "model": "gpt-3.5-turbo",
         "role": "planner",
-        "model": "gpt-3.5",
-        "instructions": "Break tasks into steps and assign to the summarizer.",
+        "instructions": "Break down tasks.",
         "tools": [],
+        "relationships": ["summarizer_{team_id}"],
+        "team_id": "{team_id}",
         "flow_order": 1,
         "qa_loop": false,
-        "team_id": "auto_team"
-      },
+        "critics": [],
+        "active": true,
+        "memory_tags": ["planning"],
+        "retry_limit": 1,
+        "description": "Plans tasks for the team."
+      }},
       ...
     ]
+
+    Output valid JSON. Include diverse roles/tools relevant to the goal.
     """
 
     response = client.chat.completions.create(
@@ -566,14 +620,13 @@ def generate_team_from_goal(goal):
             {"role": "system", "content": system_prompt},
             {"role": "user", "content": f"My goal: {goal}"}
         ],
-        temperature=0.5,
+        temperature=0.7,
         max_tokens=1200
     )
-    
+
     result_text = response.choices[0].message.content
 
     try:
-        # Extract JSON
         match = re.search(r"\[.*\]", result_text, re.DOTALL)
         if match:
             return json.loads(match.group(0))
@@ -582,6 +635,7 @@ def generate_team_from_goal(goal):
     except Exception as e:
         print("Error parsing team:", e)
         return []
+
 
 @cl.on_message
 async def handle_message(message: cl.Message):
@@ -623,7 +677,7 @@ async def handle_message(message: cl.Message):
     elif awaiting_hat_prompt:
         # Expecting description for a new hat
         # handle_prompt will set awaiting_hat_prompt to False
-        await handle_prompt(message)
+        await handle_prompt(message, ollama_llm)
 
     # --- Command Handling ---
     elif content_lower == "new blank":
@@ -642,13 +696,16 @@ async def handle_message(message: cl.Message):
     elif content_lower.startswith("edit "):
         hat_id = content.split(" ", 1)[1].strip()
         if hat_id:
-            # First, wear the hat to make it active
             await wear_hat(hat_id)
-            # Check if wearing was successful (editing_hat_id is now set)
             if cl.user_session.get("editing_hat_id") == hat_id:
-                 # Now, set the flag and prompt for JSON
                 cl.user_session.set("awaiting_json_paste", True)
-                await cl.Message(content=f"📝 OK. Paste the full updated JSON for `{hat_id}` now.").send()
+                current_hat = cl.user_session.get("current_hat")
+
+                # First Message: Instruction
+                await cl.Message(content=f"📝 OK. Paste the full updated JSON for `{hat_id}` now. Below is the current JSON for reference.").send()
+
+                # Second Message: Copyable JSON Block (Markdown Styled)
+                await cl.Message(content="```json\n" + json.dumps(current_hat, indent=2) + "\n```").send()
         else:
             await cl.Message(content="Usage: `edit <hat_id>`").send()
 
@@ -669,10 +726,7 @@ async def handle_message(message: cl.Message):
         else:
             memories = search_memory(hat_id, "", k=5)
             if memories:
-                formatted = "\n".join([
-                    f"- [{meta['timestamp']}][{meta['role']}] {doc}"
-                    for doc, meta in memories
-                ])
+                formatted = "\n".join([format_memory_entry(doc, meta) for doc, meta in memories])
                 await cl.Message(content=f"🧠 Top memories for `{hat_id}`:\n{formatted}").send()
             else:
                 await cl.Message(content=f"🧠 No memories stored for `{hat_id}` yet.").send()
@@ -829,7 +883,6 @@ async def handle_message(message: cl.Message):
         current_hat = cl.user_session.get("current_hat")
         if current_hat:
             # --- Time-Based Hat Switching ---
-            from datetime import datetime
             current_time = datetime.now().strftime("%H:%M")
             schedule = cl.user_session.get("hat_schedule", {})
 
@@ -837,6 +890,7 @@ async def handle_message(message: cl.Message):
                 scheduled_hat = schedule[current_time]
                 await wear_hat(scheduled_hat)
                 await cl.Message(content=f"🕒 Auto-switched to Hat `{scheduled_hat}` based on your schedule!").send()
+                current_hat = cl.user_session.get("current_hat")
 
 
             response_text = generate_openai_response(message.content, current_hat)
